카트 조회 구현 및  추가 로직 변경(#202)
## Summary

### 1. 장바구니 조회 구현 

### 2. 장바구니 추가 로직 수정
- 장바구니를 펀딩 전용 카트로 명확히 정의하고, 일반 상품 관련 로직 제거
- Cart / CartItem / Service / Controller 간 책임 분리 명확화
- 장바구니 아이템 추가 시 `중복 여부에 따른 결과(ADDED / UPDATED)`를 도메인에서 표현하도록 개선
---

## What's Changed



### 1. 펀딩 전용 검증 로직 명확화
- `TargetType.FUNDING_PENDING`만 장바구니에 담을 수 있도록 제한
- 펀딩 가능 여부 검증 시:
  - WishlistItem 상태(PENDING / IN_PROGRESS) 검증
  - 연결된 Product 상태(ACTIVE) 검증

- `targetId`를 내부에서는 `wishlistItemId`로 명확히 사용하여 의미 혼동 제거

### 2. addItem 결과를 도메인 의미로 표현
- 아이템 추가 결과를 `AddItemResult (CREATED / ADDED)`로 반환
- Service는 결과만 전달하고,
- Controller에서 결과에 따라 사용자 메시지 생성 (RsData 사용)

### 3. Controller 책임 정리
- 요청 → Command 변환
- Service 호출
- 도메인 결과 해석 후 API 응답 생성
- 응답 메시지 생성 로직은 Controller 내부에서 처리

+++ bc/catalog/src/main/java/app/giftify/cart/adapter/inbound/CartItemResponse.java
@@ -8,7 +8,7 @@ public record CartItemResponse(
         TargetType targetType,
         Long targetId,
         String productName,
-        int productPrice,          // 상품 원래 가격
+        long productPrice,          // 상품 원래 가격
         long contributionAmount   // 사용자가 담은 펀딩 금액
 ) {
     public static CartItemResponse from(CartItem item, Product product) {
@@ -16,7 +16,7 @@ public record CartItemResponse(
                 item.getTargetType(),
                 item.getTargetId(),
                 product.getName(),
-                product.getPrice(),
+                (long) product.getPrice(),
                 item.getAmount().amount().longValue()
         );
     }
